+++ tinker_cookbook/renderers/tml.py
@@ -23,7 +23,7 @@ from tinker_cookbook.renderers.base import (
     ToolSpec,
     TrainOnWhat,
 )
-from tinker_cookbook.renderers.tml_v0 import (
+from tinker_cookbook.renderers.tml_conversions import (
     TmlRenderInput,
     _cookbook_messages_to_sft_input,
     _messages_to_render_input,
