+++ tinker_cookbook/renderers/tml_test.py
@@ -8,7 +8,7 @@ import pytest
 import tinker
 
 from tinker_cookbook.exceptions import RendererError
-from tinker_cookbook.renderers import tml, tml_v0
+from tinker_cookbook.renderers import tml, tml_conversions, tml_v0
 from tinker_cookbook.renderers.base import (
     Message,
     ParseTermination,
@@ -191,6 +191,7 @@ def _patch_public_modules(monkeypatch: pytest.MonkeyPatch) -> None:
         return _chat_module() if name == "tml_renderers.chat" else _TinkerBridge
 
     monkeypatch.setattr(tml, "import_module", load)
+    monkeypatch.setattr(tml_conversions, "import_module", load)
     monkeypatch.setattr(tml_v0, "import_module", load)
