+++ tinker_cookbook/renderers/tml_v0_test.py
@@ -17,6 +17,7 @@ from tinker_cookbook.renderers import (
     ToolCall,
     TrainOnWhat,
     get_renderer,
+    tml_conversions,
     tml_v0,
 )
 from tinker_cookbook.supervised.data import conversation_to_datum
@@ -75,7 +76,7 @@ class _FakeTmlRenderersChat:
 def mock_tml_renderers_chat(monkeypatch: pytest.MonkeyPatch) -> type[_FakeTmlRenderersChat]:
     _FakeTmlRenderersChat.OpenAIMessage.source_messages = None
     monkeypatch.setattr(
-        tml_v0,
+        tml_conversions,
         "import_module",
         lambda module="tml_renderers": _FakeTmlRenderersChat,
     )
@@ -141,7 +142,7 @@ def test_cookbook_dicts_are_converted_with_tool_calls_for_openai_message(
         ),
     ]
 
-    rendered = cast(Any, tml_v0._messages_to_render_input(messages))
+    rendered = cast(Any, tml_conversions._messages_to_render_input(messages))
 
     assert all(isinstance(message, mock_tml_renderers_chat.OpenAIMessage) for message in rendered)
     source_messages = mock_tml_renderers_chat.OpenAIMessage.source_messages
@@ -164,7 +165,7 @@ def test_cookbook_audio_bytes_are_normalized_to_openai_input_audio(
         )
     ]
 
-    tml_v0._messages_to_render_input(messages)
+    tml_conversions._messages_to_render_input(messages)
 
     source_messages = mock_tml_renderers_chat.OpenAIMessage.source_messages
     assert source_messages is not None
@@ -197,7 +198,7 @@ def test_cookbook_audio_local_path_and_metadata_are_normalized(
         )
     ]
 
-    tml_v0._messages_to_render_input(messages)
+    tml_conversions._messages_to_render_input(messages)
 
     source_messages = mock_tml_renderers_chat.OpenAIMessage.source_messages
     assert source_messages is not None
@@ -211,14 +212,14 @@ def test_cookbook_audio_local_path_and_metadata_are_normalized(
 
 def test_cookbook_audio_non_wav_requires_complete_metadata() -> None:
     with pytest.raises(ValueError, match="must provide num_frames and sample_rate"):
-        tml_v0._audio_part_to_openai(
+        tml_conversions._audio_part_to_openai(
             AudioPart(type="audio", audio=b"mp3", format="mp3", num_frames=48_000)
         )
 
 
 def test_cookbook_audio_metadata_must_be_positive() -> None:
     with pytest.raises(ValueError, match="must be positive"):
-        tml_v0._audio_part_to_openai(
+        tml_conversions._audio_part_to_openai(
             AudioPart(
                 type="audio",
                 audio=b"mp3",
@@ -233,7 +234,7 @@ def test_cookbook_audio_remote_url_is_rejected() -> None:
     part = AudioPart(type="audio", audio="https://example.com/clip.wav")
 
     with pytest.raises(ValueError, match="does not fetch remote audio URLs"):
-        tml_v0._audio_part_to_openai(part)
+        tml_conversions._audio_part_to_openai(part)
 
 
 def test_native_tml_renderers_inputs_pass_through(
@@ -245,8 +246,8 @@ def test_native_tml_renderers_inputs_pass_through(
     ]
     message_list = mock_tml_renderers_chat.MessageList()
 
-    assert tml_v0._messages_to_render_input(cast(Any, messages)) is messages
-    assert tml_v0._messages_to_render_input(cast(Any, message_list)) is message_list
+    assert tml_conversions._messages_to_render_input(cast(Any, messages)) is messages
+    assert tml_conversions._messages_to_render_input(cast(Any, message_list)) is message_list
 
 
 def test_selective_sft_masking_sets_zero_training_metadata(
@@ -260,7 +261,10 @@ def test_selective_sft_masking_sets_zero_training_metadata(
     ]
 
     rendered = cast(
-        Any, tml_v0._cookbook_messages_to_sft_input(messages, TrainOnWhat.LAST_ASSISTANT_MESSAGE)
+        Any,
+        tml_conversions._cookbook_messages_to_sft_input(
+            messages, TrainOnWhat.LAST_ASSISTANT_MESSAGE
+        ),
     )
 
     assert rendered[1].message_metadata.training_metadata.weight == 0.0
@@ -276,7 +280,9 @@ def test_customized_sft_masking_respects_trainable_flag(
         Message(role="assistant", content="Train.", trainable=True),
     ]
 
-    rendered = cast(Any, tml_v0._cookbook_messages_to_sft_input(messages, TrainOnWhat.CUSTOMIZED))
+    rendered = cast(
+        Any, tml_conversions._cookbook_messages_to_sft_input(messages, TrainOnWhat.CUSTOMIZED)
+    )
 
     assert rendered[1].message_metadata.training_metadata.weight == 0.0
     assert rendered[2].message_metadata is None
@@ -286,7 +292,7 @@ def test_selective_sft_rejects_native_tml_renderers_inputs(
     mock_tml_renderers_chat: type[_FakeTmlRenderersChat],
 ) -> None:
     with pytest.raises(NotImplementedError, match="selective train_on_what"):
-        tml_v0._cookbook_messages_to_sft_input(
+        tml_conversions._cookbook_messages_to_sft_input(
             cast(Any, [mock_tml_renderers_chat.OpenAIMessage("assistant")]),
             TrainOnWhat.LAST_ASSISTANT_MESSAGE,
         )
